+++ compiler/semmagic.nim
@@ -9,7 +9,7 @@
 
 # This include file implements the semantic checking for magics.
 # included from sem.nim
-
+import tables
 proc semAddrArg(c: PContext; n: PNode; isUnsafeAddr = false): PNode =
   let x = semExprWithType(c, n)
   if x.kind == nkSym:
@@ -586,5 +586,13 @@ proc magicsAfterOverloadResolution(c: PContext, n: PNode,
     result = n
   of mPrivateAccess:
     result = semPrivateAccess(c, n)
+  of mDeferImport:
+    let npath = semConstExpr(c, n[1])
+    let path = pathSubs(c.config, npath.strVal, config = "") # config doesn't seem relevant here
+    let file = findModule(c.config, path, toFullPath(c.config, npath.info))
+    let fileIdx = fileInfoIdx(c.config, file)
+    if fileIdx notin c.graph.deferImports:
+      c.graph.deferImports[fileIdx] = true
+    result = newNodeIT(nkEmpty, n.info, getSysType(c.graph, n.info, tyVoid))
   else:
     result = n