+++ lib/std/importutils.nim
@@ -42,3 +42,16 @@ proc privateAccess*(t: typedesc) {.magic: "PrivateAccess".} =
     # this also works with generics
     privateAccess(Goo)
     assert Goo[float](g0: 1).g0 == 1
+
+when defined(nimHasDeferImport) and defined(nimExperimentalDeferImport):
+  proc deferImport*(path: string) {.magic: "DeferImport", compileTime.} =
+    ##[
+    This registers `path` as a module that should be imported after the main project
+    has been compiled. This helps dealing with cyclic dependencies in many cases.
+
+    `path` is interpreted using the same rules as `import "someString"`.
+    ]##
+    runnableExamples("-d:nimExperimentalDeferImport"):
+      deferImport "$lib/../tests/stdlib/mimportutils2.nim" # defines `mimportutils2_fn1`
+      proc mimportutils2_fn1(): int {.importc.}
+      assert mimportutils2_fn1() == 1